==> step_1_best_in_show.ipynb <==
This notebook converts and cleans our main source of data, the Oscars Award csv file containing all nominees and winners

#Import dependencies
import pandas as pd

#Store csv files into variables
best_in_show = "Resources/the_oscar_award.csv"

# Load in csv as a dataframe
best_in_show_df = pd.read_csv(best_in_show)
best_in_show_df

# Look at all categories and determine which are for best film
best_in_show_df["category"].unique()

# Best film category names are "OUTSTANDING MOTION PICTURE", "BEST PICTURE", "OUTSTANDING PRODUCTION",
# "OUTSTANDING MOTION PICTURE", "BEST MOTION PICTURE"
# Limit dataframe to just these categories
categories_to_filter = ["OUTSTANDING MOTION PICTURE", "BEST PICTURE", "OUTSTANDING PRODUCTION",
                        "OUTSTANDING MOTION PICTURE", "BEST MOTION PICTURE"]
best_picture = best_in_show_df[best_in_show_df["category"].isin(categories_to_filter)]
best_picture

# Save filtered results to a csv
best_picture.to_csv("Resources/best_picture.csv", index=False)
