==> step_2_OMDB_API.ipynb <==
This notebook's purpose is to run an API call to gather more data on our Best Picture nominees and winners. We also run API calls on movies anticipated to be nominated at the 2024 Oscars and for fun we've also done API calls on older movies that our group feels were overlooked and did not receive a Best Picture nomination.

from config import api_key
import requests
import json
import pandas as pd

# Read in the best_picture csv as a dataframe
best_picture = pd.read_csv("Resources/best_picture.csv")
best_picture

# Create a function that will do the OMDB API call
omdb_data = []
def omdb_api(dataframe):
    for index, row in dataframe.iterrows():
        # we need the film + year for the api call
        film = dataframe.loc[index, "film"]
        film = film.replace(" ", "+")
        year = dataframe.loc[index, "year_film"]
        try:
            # define target url
            target_url = f"http://www.omdbapi.com/?apikey={api_key}&t={film}&y={year}"
            # make api call
            response = requests.get(target_url).json()
            omdb_data.append(response)
            print("FILM FOUND")
        except:
            print(f"{film} NOT FOUND")

# Run API call on historical Best Picture dataset

# Run the API call on all movies in the best_picture dataframe 
omdb_api(best_picture)

# Save the API call results to a dataframe
omdb_data_df = pd.DataFrame(omdb_data)

# Preview the dataframe
omdb_data_df.head()

# Remove any rows where Response = False, since these movies had no data to be found
omdb_data_df = omdb_data_df.loc[omdb_data_df["Response"]=="True"]
omdb_data_df.head()

# Use the shape attribute to see how many movies had API call results
omdb_data_df.shape

# Save the API call results to a csv
omdb_data_df.to_csv("Resources/omdb_data.csv", index=False)

# Run API call on predicted 2024 Best Picture nominees

# Via an internet search, the following 5 films are some of those that the media is predicting to be nominated in 2024
potential_nominees = {
"film": ["Barbie", "Oppenheimer", "Past Lives", "Spider-Man: Across the Spider-Verse", "Air"],
"year_film": 2023}

# Create a dataframe with all these movies
potential_nominees_df = pd.DataFrame(potential_nominees)
potential_nominees_df

# Reset the API call results and run the API call on these potential nominees
omdb_data = []
omdb_api(potential_nominees_df)

# Save the API call results to a dataframe
ombd_potential_nominees_df = pd.DataFrame(omdb_data)
ombd_potential_nominees_df

# Save the API call results to a csv
ombd_potential_nominees_df.to_csv("Resources/potential_nominees.csv", index=False)

# Run API Call on historic movies previously overlooked for best picture nomination. Movies picked by group members. Doing this mainly for fun :)

# Create the data
overlooked = {
"film": ["Hocus Pocus", "Mean Girls", "Dazed & Confused", "Die Hard", "Rocky Horror Picture Show", "Princess Bride",
          "Singin' in the Rain"],
"year_film": [1993, 2004, 1993, 1988, 1975, 1987, 1952]}

# Create the dataframe
overlooked_df = pd.DataFrame(overlooked)
overlooked_df

# Run the API call
omdb_data = []
omdb_api(overlooked_df)

# Save API call results to a dataframe
ombd_overlooked_df = pd.DataFrame(omdb_data)
ombd_overlooked_df

# Save the API call results to a csv
ombd_overlooked_df.to_csv("Resources/overlooked.csv", index=False)
