==> step_5_model_in_action.ipynb <==
This notebook puts our chosen model in action and uses it to see if we can predict the winner of the Best Picture Award at the 2024 Oscars. We also have a bit of fun and see if a few movies of our choosing that were never nominated actually deserved the award. 

# Import our dependencies
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegression


# Check what unique values we have for our three Genre columns
def genre_list(dataframe):
    genres = []
    genres.extend(dataframe["Genre1"].values)
    genres.extend(dataframe["Genre2"].values)
    genres.extend(dataframe["Genre3"].values)
    global genres_list
    genres_list = list(set(genres))


# Perform one-hot encoding on Genres
def genre_encoding(dataframe):
    for g in genres_list:
        # Create a column for each genre
        dataframe[g] = 0
        # Columns will have a 0 or 1 if the movie is of the column's genre
        dataframe[g] = ((dataframe["Genre1"] == g) | (dataframe["Genre2"] == g) | (dataframe["Genre3"] == g)).astype(int)
        

def prepare_input_data(dataframe):
    # Save the movie titles to a series that will be joined to the prediction results later
    global nominees
    nominees = pd.DataFrame(dataframe["Title"])
    # Use previous functions to one-hot encode Genres
    genre_list(dataframe)
    genre_encoding(dataframe)
    dataframe.drop(columns=["Genre1", "Genre2", "Genre3"], inplace=True)
    # Run get_dummies on our Rated Column
    dataframe = pd.get_dummies(dataframe, columns=["Rated"])
    # Scaling the numeric columns
    scaled_data = StandardScaler().fit_transform(dataframe[["Runtime", "Metascore", "imdbRating", "imdbVotes", "BoxOffice"]])
    # Create a DataFrame of the scaled data
    scaled_data = pd.DataFrame(scaled_data, columns=["Runtime", "Metascore", "imdbRating", "imdbVotes", "BoxOffice"])
    # Replace the original data with the columns of information from the scaled Data
    dataframe["Runtime"] = scaled_data["Runtime"]
    dataframe["Metascore"] = scaled_data["Metascore"]
    dataframe["imdbRating"] = scaled_data["imdbRating"]
    dataframe["imdbVotes"] = scaled_data["imdbVotes"]
    dataframe["BoxOffice"] = scaled_data["BoxOffice"]
    global input_data
    input_data = dataframe.copy()


def run_prediction(dataframe):
    # Run model on our potential nominees
    predictions = classifier.predict(dataframe)
    predictions_df = pd.DataFrame({"Prediction": predictions})
    predictions_df["Prediction"] = predictions_df["Prediction"].replace({0: "It's an honor to be nominated", 1: "WINNER"})
    global oscars_prediction
    oscars_prediction = pd.concat([nominees, predictions_df], axis=1)
    return oscars_prediction


## Upload Training data and recreate the model

# Load in our training dataframe
model_training_data = pd.read_csv("Resources/model_training_data.csv")
model_training_data.head()


# Recreate the model
X = model_training_data.drop(columns = "OscarsWinner")
y = model_training_data["OscarsWinner"]
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=29)
classifier = LogisticRegression(random_state=29)
classifier.fit(X_train, y_train)


## Now we will use the model to see if we can predict what movie will win the best picture Oscar in 2024

# Load in our potential nominees dataframe
potential_nominees = pd.read_csv("Resources/potential_nominees_clean.csv")
potential_nominees


# Run the function to prepare the testing data
prepare_input_data(potential_nominees)


# Find out which columns are missing that the model was trained on
model_training_data
missing_columns = list(set(list(model_training_data.columns)) - set(list(input_data.columns)))
# Add missing columns to dataframe
input_data[missing_columns]=0
# Find out which columns are extra that our model doesn't use
extra_columns = list(set(list(input_data.columns))-set(list(model_training_data.columns)))
# Remove the extra columns
input_data.drop(columns=extra_columns, inplace=True)
# Reorder the columns so that they're in the same order as the model is trained on
input_data = input_data[list(model_training_data.columns)]
# Since this data is going to be used in place of the testing data, we drop the "OscarsWinner" column
input_data.drop("OscarsWinner",axis=1, inplace=True)
input_data


# Run the function to get prediction results
run_prediction(input_data)


## According to the model, none of these movies will win the best picture Oscar.
## Just like the 2024 Oscars, the 2024 Golden Globes haven't happened yet.
## So would the prediction change if we said that the movies win at the globes?

# Change the input_data to show the prediction for each movie if it won the Golden Globe
input_data["GlobesWinner"] = 1


# Run the function to get prediction results
run_prediction(input_data)


## According to our model, if Oppenheimer wins the Golden Globe for best picture, it will win the Oscar too!

## Now we will use model attempt #5 to see if any of our "Overlooked" movies could win an Oscar

# Load in our overlooked nominees dataframe
overlooked = pd.read_csv("Resources/overlooked_clean.csv")
overlooked


# Run the function to prepare the testing data
prepare_input_data(overlooked)


# Find out which columns are missing that the model was trained on
model_training_data
missing_columns = list(set(list(model_training_data.columns)) - set(list(input_data.columns)))
# Add missing columns to dataframe
input_data[missing_columns]=0
# Find out which columns are extra that our model doesn't use
extra_columns = list(set(list(input_data.columns))-set(list(model_training_data.columns)))
# Remove the extra columns
input_data.drop(columns=extra_columns, inplace=True)
# Reorder the columns so that they're in the same order as the model is trained on
input_data = input_data[list(model_training_data.columns)]
# Since this data is going to be used in place of the testing data, we drop the "OscarsWinner" column
input_data.drop("OscarsWinner",axis=1, inplace=True)
input_data


# Run the function to get prediction results
run_prediction(input_data)


## According to our model, Singin' in the Rain should have won an Oscar for best picture.
